requirements.txt
boto==2.49.0
Django==3.2.7
django_phonenumber_field==6.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.0
gunicorn==20.1.0
whitenoise==6.2.0